// gr-checks/types.ts
/**
 * @license
 * Copyright (C) 2021 The Android Open Source Project
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface Check {
  state: string;
  checker_name: string;
  checker_description: string;
  checker_uuid: string;
  message: string;
  url?: string;
  started?: number;
  finished?: number;
}

export interface Checker {
  repository?: string;
  blocking?: string[];
  query?: string;
  url?: string;
  description?: string;
  name: string;
  uuid: string;
  status: string;
}

export interface ProjectInfo {

}